Fixed chunks only rendering in one direction

+++ src/Input.h
@@ -8,7 +8,7 @@ public:
 	inline static bool IsMouseButtonPressed(int button) { return s_Instance->IsMouseButtonPressedImpl(button); }
 	inline static float GetMouseX() { return s_Instance->GetMouseXImpl(); }
 	inline static float GetMouseY() { return s_Instance->GetMouseYImpl(); }
-	inline static std::pair<float, float> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
+	inline static std::pair<double, double> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
 protected:
 	virtual bool IsKeyPressedImpl(int keycode) = 0;
